backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.23.6
botocore==1.26.6
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-cors-headers==3.12.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
django-filter==21.1
django-oauth-toolkit==2.0.0
drf-access-policy==1.1.1
editdistance==0.6.0
elastic-transport==8.1.2
elasticsearch==7.14.0
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
markdown==3.3.7
mongoengine==0.23.1
numpy==1.23.1
opencv-python==4.6.0.66
psycopg2-binary==2.9.3
PyJWT==2.4.0
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.3
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
